Testing with pgmpy and PyAgrum
PyAgrum DOES have a clone method. However, pruning nodes result in memory leakage, which could be dangerous. On the other hand pgmpy seems much more robust, but the inference is slower and we don't have the concept of "inference engine" (which allows to compute the marginal of every single node given an evidence using a single propagation)